backend/requirement.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-multipart==0.0.6
pillow==10.0.0
numpy==1.26.0
tensorflow==2.14.0
torch==2.2.0
torchvision==0.17.0
# adjust versions to match your environment
